# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenPanel API
  description: Self-hosted manga/comic reader server
  version: 0.1.1
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6515
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/auth/login
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        profile: { $ref: '#/components/schemas/AuthProfile' }
    AuthProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        is_admin: { type: boolean }
...
    SeriesItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        book_count: { type: integer }
        book_type: { type: string }
        year: { type: integer, nullable: true }
        anilist_cover_url: { type: string, nullable: true }
        anilist_score: { type: integer, nullable: true }
    AllSeriesResponse:
      type: object
      properties:
        series:
          type: array
          items: { $ref: '#/components/schemas/SeriesItem' }
        total: { type: integer }
    SeriesListResponse:
      type: object
      properties:
        series:
          type: array
          items: { $ref: '#/components/schemas/SeriesItem' }
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register initial admin account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
...
  /api/libraries/{library_id}/series:
    get:
      tags: [Libraries]
      summary: List series in library
      parameters:
        - name: library_id
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: per_page
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesListResponse' }

  # ── Series ──
  /api/series:
    get:
      tags: [Series]
      summary: All series (filterable)
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: per_page
          in: query
          schema: { type: integer }
        - name: sort
          in: query
          schema: { type: string, enum: [name, updated_at, created_at, year, score] }
        - name: sort_dir
          in: query
          schema: { type: string, enum: [asc, desc] }
        - name: genre
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: year
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllSeriesResponse' }
...
  /api/series/recently-added:
    get:
      tags: [Series]
      summary: Recently added series
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SeriesItem' }
  /api/series/recently-updated:
    get:
      tags: [Series]
      summary: Recently updated series
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SeriesItem' }
  /api/series/{series_id}/books:
    get:
      tags: [Series]
      summary: Books in series
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BooksListResponse' }
  /api/series/{series_id}/chapters:
    get:
      tags: [Series]
      summary: Chapters across all books
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  series_id: { type: string }
                  total_chapters: { type: integer }
                  chapters:
                    type: array
                    items: { $ref: '#/components/schemas/SeriesChapter' }
  /api/series/{series_id}/rescan:
    post:
      tags: [Series]
      summary: Rescan series
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
                  books_scanned: { type: integer }
  /api/series/{series_id}/metadata:
    get:
      tags: [Series]
      summary: Get series metadata
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesMetadata' }
    put:
      tags: [Series]
      summary: Set AniList ID
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [anilist_id]
              properties:
                anilist_id: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesMetadata' }
    delete:
      tags: [Series]
      summary: Clear metadata
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesMetadata' }
  /api/series/{series_id}/metadata/refresh:
    post:
      tags: [Series]
      summary: Refresh metadata from AniList
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesMetadata' }

  # ── Books ──
  /api/books/{book_id}:
    get:
      tags: [Books]
      summary: Book detail
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BookDetailResponse' }
  /api/books/{book_id}/chapters:
    get:
      tags: [Books]
      summary: Chapters within a book
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  book_id: { type: string }
                  chapters:
                    type: array
                    items: { $ref: '#/components/schemas/BookChapter' }

  # ── Reader ──
  /api/books/{book_id}/pages/{page_num}:
    get:
      tags: [Reader]
      summary: Serve page image
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
        - name: page_num
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Image data
          content:
            image/*:
              schema:
                type: string
                format: binary
  /api/books/{book_id}/download:
    get:
      tags: [Reader]
      summary: Download book archive
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Archive file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /api/books/{book_id}/manifest:
    get:
      tags: [Reader]
      summary: Page manifest
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageManifestResponse' }
  /api/books/{book_id}/thumbnail:
    get:
      tags: [Reader]
      summary: Book thumbnail
      parameters:
        - name: book_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Thumbnail image
          content:
            image/webp:
              schema:
                type: string
                format: binary
  /api/series/{series_id}/thumbnail:
    get:
      tags: [Reader]
      summary: Series cover thumbnail
      parameters:
        - name: series_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Thumbnail image
          content:
            image/webp:
              schema:
                type: string
                format: binary

  # ── Progress ──
  /api/progress:
    get:
      tags: [Progress]
      summary: Get progress for a book
      parameters:
        - name: book_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProgressResponse' }
    put:
      tags: [Progress]
      summary: Update reading progress
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateProgressRequest' }
      responses:
        '200': { description: OK }
  /api/progress/batch:
    get:
      tags: [Progress]
      summary: Batch progress
      parameters:
        - name: book_ids
          in: query
          required: true
          schema: { type: string }
          description: Comma-separated book IDs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BatchProgressResponse' }
  /api/progress/bulk-mark:
    post:
      tags: [Progress]
      summary: Bulk mark as read/unread
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BulkMarkRequest' }
      responses:
        '200': { description: OK }
...
  /api/series-continue:
    get:
      tags: [Progress]
      summary: Continue point for a series
      parameters:
        - name: series_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SeriesContinueResponse' }

  # ── Bookmarks ──
  /api/bookmarks:
    get:
      tags: [Bookmarks]
      summary: List bookmarks
      parameters:
        - name: book_id
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Bookmark' }
    post:
      tags: [Bookmarks]
      summary: Create bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBookmarkRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Bookmark' }
  /api/bookmarks/{bookmark_id}:
    delete:
      tags: [Bookmarks]
      summary: Delete bookmark
      parameters:
        - name: bookmark_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }

  # ── Collections ──
...